+++ Cargo.toml
@@ -61,6 +61,7 @@ which = "6.0.1"
 
 [features]
 git = ["dep:git2"]
+ssh = []
 
 # The profile that 'cargo dist' will build with
 [profile.dist]
