+++ src/git/commands.rs
@@ -1,5 +1,3 @@
-use std::path::PathBuf;
-
 use clap::{Args, Subcommand};
 
 /// Git helpers
@@ -14,18 +12,4 @@ pub struct GitCommand {
 pub enum GitSubCommands {
     /// Setup your local git config to use ssh and a seperate .gitconfig for your work projects
     Setup,
-    /// Create .ssh/config entries for your tailscale hosts.
-    TailscaleAlias {
-        /// username for the ssh connection
-        #[arg(short = 'u', long)]
-        user: String,
-        /// ssh identity file path
-        #[arg(short = 'i', long)]
-        identity_file: PathBuf,
-        /// ssh identity file path
-        #[arg(short = 'f', long)]
-        tag_filter: Option<String>,
-        #[arg(short = 'p', long)]
-        hostname_prefix: Option<String>,
-    },
 }
