+++ src/ssh/mod.rs
@@ -1,0 +1,5 @@
+mod commands;
+mod ssh;
+
+pub use commands::SshCommands;
+pub use SshPlugin;
